+++ ai-compile/vscode-extension/src/codeLensProvider.ts
@@ -1,6 +1,6 @@
 import * as vscode from "vscode";
-import * as t from "./types";
 import { docStore } from "./extension";
+import * as pymacer from "./pymacer";
 
 export class CodelensProvider implements vscode.CodeLensProvider {
   private codeLenses: vscode.CodeLens[] = [];
@@ -28,8 +28,7 @@ export class CodelensProvider implements vscode.CodeLensProvider {
     ) {
       this.codeLenses = [];
       const filePath = document.uri.fsPath;
-      const fixes: t.Fix = docStore.get(filePath)
-        ?.fixes;
+      const fixes: pymacer.Fix = docStore.get(filePath)?.fixes;
       fixes?.forEach((fix) => {
         const position = new vscode.Position(fix.lineNo, 0);
         const range = document.getWordRangeAtPosition(
