+++ ai-compile/vscode-extension/src/extension.ts
@@ -1,90 +1,18 @@
 // The module 'vscode' contains the VS Code extensibility API
 // Import the module and reference it with the alias vscode in your code below
 import * as vscode from "vscode";
+
 import * as path from "path";
-import { compile, getFix } from "./util";
 import { CodelensProvider } from "./codeLensProvider";
 import { Decorator } from "./decorator";
 
-import * as t from "./types";
-import * as c from "./constants";
+import * as pymacer from "./pymacer";
 
 let disposables: vscode.Disposable[] = [];
 let eventDisposables: vscode.Disposable[] = [];
 
 // file-wise history store
-export let docStore: Map<string, t.DocumentStore>;
-
-async function compileAndGetFixHelper(): Promise<t.Fix> {
-  // TODO: What if user changes tab immediately? - activeEditor changes - cancellation token?
-  //* Or find an alternate way to get fullText of document and simply pass document along
-  const activeEditor = vscode.window.activeTextEditor;
-
-  if (activeEditor !== undefined) {
-    const document = activeEditor.document;
-    const cursorPosition = activeEditor.selection.active;
-    const filePath = document.uri.fsPath;
-    let result: t.Fix = undefined;
-
-    console.log(`Compiling ${filePath}`);
-    const compiled = await compile(filePath);
-
-    if (!compiled) {
-      console.log("Syntax Error -> Preparing and Sending data...");
-
-      const data: t.Payload = {
-        // srcCode: getDocumentText( document ),
-        source: document.getText(),
-        lastEditLine: cursorPosition.line,
-      };
-      const payload: t.Payload = JSON.parse(JSON.stringify(data));
-
-      if (c.baseURL === undefined) {
-        console.log("Web Server path invalid");
-        return undefined;
-      } else {
-        result = await getFix(c.baseURL!, payload);
-        if (result !== undefined) {
-          console.log("Reply from Server:");
-        }
-      }
-    }
-
-    return result;
-  }
-}
-
-async function compileAndGetFix(
-  document: vscode.TextDocument,
-  docStore: Map<string, t.DocumentStore>
-): Promise<t.Fix> {
-  const filePath = document.uri.fsPath;
-  let fixes: t.Fix = undefined;
-  let docHistory = docStore.get(filePath);
-
-  // TODO: What if the document was left in an inconsistent state previously
-  // * Can handle this case in the document open/ close event?
-  // compile if file has either not been examined previously (in the history of the extension) ||
-  // it has been modified
-  let compileFlag = docHistory === undefined || document.isDirty;
-
-  // docHistory = docHistory?? new Document( filePath, undefined );
-  // docHistory = docHistory?? {filePath: filePath, fixes: fixes};
-  docHistory = docHistory ?? { filePath: filePath, fixes: fixes };
-
-  if (compileFlag) {
-    fixes = await compileAndGetFixHelper();
-    docHistory.fixes = fixes;
-    docStore.set(filePath, docHistory);
-  } else {
-    fixes = docHistory.fixes;
-    if (fixes !== undefined) {
-      console.log("Saved fixes from History:");
-    }
-  }
-
-  return fixes;
-}
+export let docStore: Map<string, pymacer.DocumentStore>;
 
 export function activate(context: vscode.ExtensionContext) {
   console.log("Extension 'python-hints' is now active!");
@@ -114,7 +42,7 @@ export function activate(context: vscode.ExtensionContext) {
             activeEditor.document.save();
           } else {
             // document.save() doesn't trigger if document isn't dirty
-            const fixes = await compileAndGetFix(
+            const fixes = await pymacer.compileAndGetFix(
               activeEditor.document,
               docStore
             );
@@ -166,7 +94,10 @@ export function activate(context: vscode.ExtensionContext) {
         saveEvent.reason === vscode.TextDocumentSaveReason.Manual &&
         path.basename(saveEvent.document.uri.fsPath.split(".")[1]) === "py"
       ) {
-        const fixes = await compileAndGetFix(saveEvent.document, docStore);
+        const fixes = await pymacer.compileAndGetFix(
+          saveEvent.document,
+          docStore
+        );
 
         console.log(fixes);
         decorator.updateDecorations();
